requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
jupyter==1.0.0
notebook==6.5.4
torch==2.0.1
tensorflow==2.13.0
pytest==7.4.0

# Web scraping libraries
requests==2.31.0
selenium==4.11.2
playwright==1.38.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0

# Data storage
pyarrow==14.0.1
fastparquet==2024.2.0

# Progress bars
tqdm==4.66.1

pycoingecko==3.2.0
polars==1.29.0